requirements.txt
asgiref==3.7.2
awscli==1.32.5
boto3==1.34.3
botocore==1.34.5
colorama==0.4.4
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
docutils==0.16
jmespath==1.0.1
Markdown==3.5.1
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
